[domain] move Filter into module
